pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "project"
version = "0.1.0"
description = "Project starter"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "ppofluxus", email = "ppofluxus@gmail.com" }]
dependencies = []

[project.optional-dependencies]
# Heavy ML deps are optional; install with:
#   pip install -e .[timeseries]
#   pip install -e .[vision]
timeseries = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "matplotlib>=3.7",
    "tensorflow>=2.12",
]
vision = [
    "numpy>=1.24",
    "pandas>=2.0",
    "pillow>=10.0",
    "scikit-learn>=1.3",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "tensorflow>=2.12",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
